historico dos pedido

+++ empresa/forms.py
@@ -10,7 +10,7 @@ class PedidosForm(forms.ModelForm):
         fields = '__all__'
     empresa = forms.ModelChoiceField(queryset=Empresa.objects.all(), initial=Empresa.objects.get(pk=1))
     quantidade = forms.IntegerField(required=False)
-    valor = forms.IntegerField(required=False)
+    valor = forms.FloatField(required=False)
     codigo = forms.CharField(required=False)
     data_entrega = forms.DateField(required=False, widget=forms.DateInput(attrs={'type': 'date'}))
